requirements.txt
python-dotenv
open-interpreter
pandas
odfpy